Refactoring for Mac dev
* Moved combustion logic into its own module
* Added a Dockerfile so I can code on my Mac
* Added a simple Mac implementation with conditional compilation
