+++ Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-bluer = { version = "0.16.1", features = ["full"] }
+anyhow = "1.0.77"
 bytes = "1"
 env_logger = "0.10.1"
 futures = "0.3.29"
@@ -14,6 +14,9 @@ hyper = { version = "1", features = ["full"] }
 http-body-util = "0.1"
 hyper-util = { version = "0.1", features = ["full"] }
 log = "0.4.20"
-modular-bitfield = "0.11.2"
 tokio = { version = "1", features = ["full"] }
 tokio-stream = { version = "0.1", optional = true }
+
+[target.'cfg(target_os = "linux")'.dependencies]
+bluer = { version = "0.16.1", features = ["full"] }
+modular-bitfield = "0.11.2"
